CA-135649: Add additional HBA information to sr-probe and sr-create
There currently is no easy way to gather the basic information about HBAs, that
are required for provisioning external storage.

This commit adds some basic information about the FC and iSCSI HBAs to the
HBA-specific output of "xe sr-probe type=LVMoHBA" and "xe sr-create
type=LVMoHBA" to address this gap.

Signed-off-by: Robert Breker <robert.breker@citrix.com>
Reviewed-by: Chandrika Srinivasan <chandrika.srinivasan@citrix.com>