CA-141409: Default to using NFS version 3 again
We were unintentionally using nfsversion 4 on trunk.
This was because the command mount.nfs defaults to using nfsversion 4 instead
of version 3 in our environment on trunk, despite it's documentation claiming
otherwise.
This commit specifies the nfsversion additionally in the mount options, to make
sure that we default to nfsversion 3 again, and that the default setting can
be switched intentionally.

Signed-off-by: Robert Breker <robert.breker@citrix.com>

[Resolved conflicts in drivers/nfs.py]
Signed-off-by: Siddharth Vinothkumar <siddharth.vinothkumar@citrix.com>