CP-10986: Return read caching status on VDI.attach
Returning the status of read caching during VDI Attach is a bit tricky.
This is because the SM only works out whether or not O_DIRECT should be
used during VDI Activate (which happens after attach). This is determined by a
number of elements: licensing, override parameters or VDI Type, etc.

This patch isolates this check in a separate (private) method and
caches it. Both VDI Attach and Activate use this method to work out whether
tapdisk will be enabling read caching or not.

We also add this information to the return struct in VDI.attach as
requested by the ticket.

Signed-off-by: Felipe Franciosi <felipe@paradoxo.org>
Signed-off-by: Siddharth Vinothkumar <siddharth.vinothkumar@citrix.com>